change output data type

+++ main.py
@@ -47,7 +47,7 @@ def predict(image_path = "./image.tif", save_path = "./prediction.tif", weight_p
 
     predict_windows(pathTif = image_path, pathSave = save_path, predictor = predictor, preprocess = preprocess,
                     window_size = 512, input_dim = input_dim, predict_dim = predict_dim,
-                    output_type = "float32", batch_size = batch_size)
+                    output_type = "int8", batch_size = batch_size)
 
 if __name__ == "__main__":
     main_args = get_main_args()
